startup-rag/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.0
pydantic==2.10.6
python-multipart==0.0.21
requests==2.31.0
beautifulsoup4==4.12.2
lxml
langdetect==1.0.9